feat: Move to builder directory to output data
